// src/types.ts
export interface Location {
  BIKE_PATH: string;
  BLFTP_PRK: string;
  BLFTP_TRLS: string;
  BLUFF: string;
  BOATING: string;
  BT_FACIL_TYPE: string;
  Bch_whlchr: string;
  CAMPGROUND: string;
  COUNTY: string;
  CountyNum: number;
  DISTRICT: string;
  DOG_FRIENDLY: string;
  DSABLDACSS: string;
  DUNES: string;
  DescriptionMobileWeb: string;
  EZ4STROLLERS: string;
  FEE: string;
  FISHING: string;
  GEOGR_AREA: string;
  ID: number;
  LATITUDE: number;
  LIST_ORDER: string;
  LONGITUDE: number;
  LocationMobileWeb: string;
  NameMobileWeb: string;
  PARKING: string;
  PCNC_AREA: string;
  PHONE_NMBR: string;
  PTH_BEACH: string;
  Photo_1: string;
  Photo_2: string;
  Photo_3: string;
  Photo_4: string;
  RESTROOMS: string;
  RKY_SHORE: string;
  SNDY_BEACH: string;
  STRS_BEACH: string;
  TIDEPOOL: string;
  VISTOR_CTR: string;
  VOLLEYBALL: string;
  WLDLFE_VWG: string;
  photos: string[] 
}